Moved lots of test files into sub-package folders corresponding to tested modules
Fixing up paths and removing unneeded imports and other cruft.
Merged a few tests into single files - test_entity.py and test_timeline.py

+++ tools/toollib/import_analyzer.py
@@ -135,8 +135,7 @@ def get_extras_from_setup(
 
     neut_setup_py = Path(package_root) / "neut_setup.py"
     try:
-        with open(neut_setup_py, "w") as f_handle:
-            f_handle.writelines(setup_txt)
+        neut_setup_py.write_text(setup_txt)
 
         setup_mod = import_module("neut_setup")
         extras = getattr(setup_mod, "EXTRAS").get(extra)
@@ -147,7 +146,11 @@ def get_extras_from_setup(
             )
         return sorted(list(set(extras)), key=str.casefold)
     finally:
-        neut_setup_py.unlink()
+        try:
+            if neut_setup_py.is_file():
+                neut_setup_py.unlink()
+        except PermissionError:
+            print(f"could not remove temp file {neut_setup_py}")
 
 
 def _extract_pkg_specs(pkg_specs: List[str]):
